Add SuperSU to Screw'd [2/2]
Thanks to chinfire's "How to SU" guide,
I've managed to get SuperSU zip flashing on every
rom flash and it can easily be updated in
vendor_aicp

Also Settings part merged in from OmniRom

Screw'd:
Have to say big thanks to LorDClockaN for this.
I've been trying to figure this out for a while
and he hit it on the nail.

PS2: Don't unmount /system, it gets unmounted
     from within superu.zip flash script

Change-Id: I0f57f8c2b8733021c3c0b9a5fd3c343ae565a959
Signed-off-by: LorDClockaN <davor@losinj.com>